requirements.txt

https://github.com/KurimuzonAkuma/pyrogram/archive/dev.zip
TgCrypto
requests
python-dotenv
